+++ app/components/idea-card.js
@@ -4,6 +4,10 @@ import { action } from '@ember/object';
 export default class IdeaCardComponent extends Component {
   @action
   delete() {
+    if (!confirm('Are you sure you want to delete this idea?')) {
+      return;
+    }
+
     this.args.idea.destroyRecord().catch(error => {
       console.error(error);
       alert(
